third_party/blink/web_tests/virtual/single-import-map/external/wpt/import-maps/acquiring/modulepreload-expected.txt

This is a testharness.js-based test.
[FAIL] After <link rel=modulepreload> import maps should work fine
  assert_array_equals: expected property 0 to be "log:B" but got "log:A" (expected array ["log:B"] got ["log:A"])
Harness: the test ran to completion.
